The overlay directory contains user mode tools for the EM9010 overlay processor kernel driver. All communication with the driver is handled by a simple overlay API (overlay.h / overlay.c). The API contains functions for setting window size/position, video mode geometry calibration, color-keying calibration and saving/retreiving of the calibration parameters.
I've modified Edward Salley's nice GTK+ overlay viewer "dxr3view" to
(thanks Edward!) use the overlay API instead of console.pl. My changes
are very crude.
The resolution is for instance hard-coded to 1280x1024 and I wasn't able
to change the background color to the chroma-keying color so for now
the overlay feature is not used.

There is a calibration tool named "autocal" which can calibrate
the EM9010 chip to the used video mode automagically. The code is
based on the calibration tool that comes with Sigma Designs EM8400
linux driver.

It also measures
how the the colors of the video card are related to the color levels
seen by the EM9010 unit. Actually this calibration is more
accurate than the way windows driver does it. The relationship is
linear and the coefficients are calculated using a least-squares
algorithm. "autocal" can also write matlab data files which can be
viewed by plotcal.m.

Henrik Johansson