Generate em8300_version.h with a new script external to the Makefile.
Building from a dist tarball could fail if "hg id -i" sent anything to
stdout (this happens with old versions of mercurial if no repository
root directory is found) because a broken em8300_version.h is generated.

Generating em8300_version.sh with a script that is not exported in the
dist tarball solves the problem, since, because of the lack of this
script, it won't even try to run "hg id -i".

Thanks to Ville Salmela who reported the problem.